+++ packages/core/src/defaults/values/box-model-values.ts
@@ -1,31 +1,5 @@
 import { createValues } from "@/helpers/create-values";
-import { dimension, height, margin, padding, width } from "../variables";
-
-export const dimensionValues: Record<string, string> = {
-	...createValues({
-		base: dimension,
-		unit: "rem",
-		min: 0,
-		max: 100,
-	}),
-	auto: "auto",
-	dvh: "100dvh",
-	dvw: "100dvw",
-	vh: "100vh",
-	vw: "100vw",
-	fc: "fit-content",
-	full: "100%",
-	half: "50%",
-	max: "max-content",
-	min: "min-content",
-	px: "1px",
-	s: "stretch",
-	sm: "40rem",
-	md: "48rem",
-	lg: "64rem",
-	xl: "80rem",
-	xxl: "96rem",
-};
+import { height, margin, padding, width } from "../variables";
 
 export const heightValues: Record<string, string> = {
 	...createValues({
@@ -39,6 +13,14 @@ export const heightValues: Record<string, string> = {
 	dvw: "100dvw",
 	vh: "100vh",
 	vw: "100vw",
+	vi: "100vi",
+	vb: "100vb",
+	svh: "100svh",
+	svw: "100svw",
+	lvh: "100lvh",
+	lvw: "100lvw",
+	vmin: "100vmin",
+	vmax: "100vmax",
 	fc: "fit-content",
 	full: "100%",
 	half: "50%",
@@ -65,6 +47,14 @@ export const widthValues: Record<string, string> = {
 	dvw: "100dvw",
 	vh: "100vh",
 	vw: "100vw",
+	vi: "100vi",
+	vb: "100vb",
+	svh: "100svh",
+	svw: "100svw",
+	lvh: "100lvh",
+	lvw: "100lvw",
+	vmin: "100vmin",
+	vmax: "100vmax",
 	fc: "fit-content",
 	full: "100%",
 	half: "50%",